split users packages into standard and admin
