+++ gwcli/tree/admin/admin.go
@@ -13,7 +13,7 @@ import (
 	"github.com/gravwell/gravwell/v4/gwcli/connection"
 	ft "github.com/gravwell/gravwell/v4/gwcli/stylesheet/flagtext"
 	"github.com/gravwell/gravwell/v4/gwcli/tree/admin/groups"
-	"github.com/gravwell/gravwell/v4/gwcli/tree/admin/users"
+	admin_users "github.com/gravwell/gravwell/v4/gwcli/tree/admin/users"
 	"github.com/gravwell/gravwell/v4/gwcli/utilities/scaffold"
 	"github.com/gravwell/gravwell/v4/gwcli/utilities/treeutils"
 	"github.com/spf13/cobra"
@@ -30,7 +30,7 @@ func NewNav() *cobra.Command {
 	return treeutils.GenerateNav(use, short, long, []string{"administrator"},
 		[]*cobra.Command{
 			groups.NewNav(),
-			users.NewNav(),
+			admin_users.NewNav(),
 		},
 		[]action.Pair{
 			cleanup(),
