+++ gwcli/tree/admin/users/lock.go
@@ -1,4 +1,4 @@
-package users
+package admin_users
 
 import (
 	"fmt"
